Move files from `SAGESharp.SLB.Internal` into `SAGESharp.SLB` namespace
These files were created with the concepts of Java's packages in mind
but these are C#'s namespaces and the way to think about them is
different, so this commit changes the extension classes where they
should be.

If these extensions should be needed elsewhere they can always be moved
to the parent namespace.
