+++ SAGESharp/SAGESharp.csproj
@@ -52,10 +52,10 @@
     <Compile Include="INIConfig.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SLB\ASCIIExtensions.cs" />
-    <Compile Include="SLB\Internal\EqualityUtils.cs" />
+    <Compile Include="SLB\EqualityUtils.cs" />
     <Compile Include="SLB\ISLBBinaryReader.cs" />
     <Compile Include="SLB\Identifier.cs" />
-    <Compile Include="SLB\Internal\IdentifierBinaryReader.cs" />
+    <Compile Include="SLB\IdentifierBinaryReader.cs" />
     <Compile Include="SLB\Level\Conversation\Character.cs" />
     <Compile Include="SLB\Level\Conversation\LineSide.cs" />
     <Compile Include="SLB\Level\Conversation\Frame.cs" />
