+++ SAGESharp/SLB/Level/Conversation/Character.cs
@@ -1,5 +1,4 @@
-﻿using SAGESharp.SLB.Internal;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
