+++ SAGESharp/SLB/Level/Conversation/Info.cs
@@ -1,5 +1,4 @@
-﻿using SAGESharp.SLB.Internal;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
