+++ SAGESharpTests/SAGESharpTests.csproj
@@ -62,10 +62,10 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SLB\ASCIIExtensionsTests.cs" />
+    <Compile Include="SLB\EqualityUtilsTests.cs" />
     <Compile Include="SLB\IdentifierEqualityTests.cs" />
     <Compile Include="SLB\IdentifierTests.cs" />
-    <Compile Include="SLB\Internal\EqualityUtilsTests.cs" />
-    <Compile Include="SLB\Internal\IdentifierBinaryReaderTests.cs" />
+    <Compile Include="SLB\IdentifierBinaryReaderTests.cs" />
     <Compile Include="SLB\Level\Conversation\CharacterEqualityTests.cs" />
     <Compile Include="SLB\Level\Conversation\Defaults.cs" />
     <Compile Include="SLB\Level\Conversation\FrameEqualityTests.cs" />
